[REFACTOR/HM-101] 선택형 타임테이블 히트맵 타입 수정
